moved PicSearchNavHost file
