+++ app/src/main/java/com/alexeybondarenko/picsearch/ui/utils/navigation/PicSearchNavHost.kt
@@ -1,4 +1,4 @@
-package com.alexeybondarenko.picsearch.ui.main
+package com.alexeybondarenko.picsearch.ui.utils.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -7,9 +7,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.alexeybondarenko.picsearch.ui.imagesearch.ImageSearchRoute
 import com.alexeybondarenko.picsearch.ui.savedimages.SavedImagesRoute
-import com.alexeybondarenko.picsearch.ui.savedimages.SavedImagesScreen
 import com.alexeybondarenko.picsearch.ui.settings.SettingsScreen
-import com.alexeybondarenko.picsearch.ui.utils.navigation.Routes
 
 @Composable
 fun PicSearchNavHost(
